Depend on `actionpack` and `railties`, not `rails`
`rails` is a meta gem that pulls the Rails default stack together, and it shouldn't
be a dependency on gems that might be used in apps with adapted stacks
(apps without ActiveRecord, for instance).

Note that the `actionpack` dependency is to ensure backwards compatibility with
Rails 3.2, where `actionview` is not on a gem of its own, but part of `actionpack`.